+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/ImportModuleDynamicallyCallback.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -54,7 +54,7 @@ public interface ImportModuleDynamicallyCallback {
      *
      * @param realm The caller realm.
      * @param referrer Script or Module that calls import().
-     * @param moduleRequest The name of the module to be imported and its assertions.
+     * @param moduleRequest The name of the module to be imported and its attributes.
      * @return a promise from the module loader that will be resolved with the module namespace
      *         object when the module is loaded successfully or rejected on error.
      */