+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/JSContextOptions.java
@@ -685,11 +685,6 @@ public final class JSContextOptions {
     public static final OptionKey<Boolean> IMPORT_ATTRIBUTES = new OptionKey<>(false);
     @CompilationFinal private boolean importAttributes;
 
-    public static final String IMPORT_ASSERTIONS_NAME = JS_OPTION_PREFIX + "import-assertions";
-    @Option(name = IMPORT_ASSERTIONS_NAME, category = OptionCategory.USER, help = "Enable legacy import assertions") //
-    public static final OptionKey<Boolean> IMPORT_ASSERTIONS = new OptionKey<>(false);
-    @CompilationFinal private boolean importAssertions;
-
     public static final String JSON_MODULES_NAME = JS_OPTION_PREFIX + "json-modules";
     @Option(name = JSON_MODULES_NAME, category = OptionCategory.USER, help = "Enable loading of json modules") //
     public static final OptionKey<Boolean> JSON_MODULES = new OptionKey<>(false);
@@ -856,7 +851,6 @@ public final class JSContextOptions {
         this.operatorOverloading = readBooleanOption(OPERATOR_OVERLOADING);
         this.errorCause = readBooleanOption(ERROR_CAUSE, JSConfig.ECMAScript2022);
         this.importAttributes = readBooleanOption(IMPORT_ATTRIBUTES, JSConfig.ECMAScript2025);
-        this.importAssertions = readBooleanOption(IMPORT_ASSERTIONS);
         this.jsonModules = readBooleanOption(JSON_MODULES, JSConfig.ECMAScript2025);
         this.sourcePhaseImports = SOURCE_PHASE_IMPORTS.hasBeenSet(optionValues) ? readBooleanOption(SOURCE_PHASE_IMPORTS) : webAssembly;
         this.wasmBigInt = readBooleanOption(WASM_BIG_INT);
@@ -1302,10 +1296,6 @@ public final class JSContextOptions {
 
     public boolean isImportAttributes() {
         return importAttributes;
-    }
-
-    public boolean isImportAssertions() {
-        return importAssertions;
     }
 
     public boolean isJsonModules() {