+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/JSParserOptions.java
@@ -59,7 +59,6 @@ public record JSParserOptions(boolean strict,
                 boolean allowBigInt,
                 boolean classFields,
                 boolean importAttributes,
-                boolean importAssertions,
                 boolean sourcePhaseImports,
                 boolean privateFieldsIn,
                 boolean topLevelAwait,
@@ -79,20 +78,19 @@ public record JSParserOptions(boolean strict,
         boolean allowBigInt = options.bigInt();
         boolean classFields = options.classFields();
         boolean importAttributes = options.importAttributes();
-        boolean importAssertions = options.importAssertions();
         boolean sourcePhaseImports = options.sourcePhaseImports();
         boolean privateFieldsIn = options.privateFieldsIn();
         boolean topLevelAwait = options.topLevelAwait();
         boolean explicitResourceManagement = options.explicitResourceManagement();
         boolean v8Intrinsics = options.v8Intrinsics();
         return new JSParserOptions(strict, scripting, shebang, ecmaScriptVersion, syntaxExtensions, constAsVar, functionStatementError, emptyStatements, annexB, allowBigInt,
-                        classFields, importAttributes, importAssertions, sourcePhaseImports, privateFieldsIn, topLevelAwait, explicitResourceManagement, v8Intrinsics);
+                        classFields, importAttributes, sourcePhaseImports, privateFieldsIn, topLevelAwait, explicitResourceManagement, v8Intrinsics);
     }
 
     public JSParserOptions withStrict(@SuppressWarnings("hiding") boolean strict) {
         if (strict != this.strict) {
             return new JSParserOptions(strict, scripting, shebang, ecmaScriptVersion, syntaxExtensions, constAsVar, functionStatementError, emptyStatements, annexB, allowBigInt,
-                            classFields, importAttributes, importAssertions, sourcePhaseImports, privateFieldsIn, topLevelAwait, explicitResourceManagement, v8Intrinsics);
+                            classFields, importAttributes, sourcePhaseImports, privateFieldsIn, topLevelAwait, explicitResourceManagement, v8Intrinsics);
         }
         return this;
     }